src/cable_detection_picking/requirements.txt
albumentations==1.3.0
cityscapesScripts==2.2.1
depthai==2.26.0.0
ffmpegcv==0.2.6
imagecorruptions==1.1.2
imageio==2.9.0
instaboostfast==0.1.2
matplotlib==3.5.3
memory_profiler==0.61.0
mmcv_full==1.6.0
model_archiver==1.0.3
multiprocess==0.70.14
networkx==2.6.3
numpy==1.21.6
onnxruntime==1.13.1
onnxsim==0.4.10
open3d==0.16.0
opencv_contrib_python==4.5.5.62
opencv_python==4.7.0.72
opencv_python_headless==4.9.0.80
pcl==0.0.0.post1
Pillow==9.3.0
psutil==5.9.4
pycocotools==2.0.6
pyserial==3.5
pysoem==1.1.8
PyYAML==6.0.1
requests==2.28.1
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools==65.6.3
shapely==2.0.0
six==1.16.0
tensorrt==8.5.1.7
terminaltables==3.1.10
torch==1.12.0
torchvision==0.13.0
ts==0.5.1
wandb==0.13.5
